Move source tree utils into their own files
