+++ src/utils/sources-tree/addToTree.js
@@ -1,0 +1,143 @@
+import { isPretty } from "../source";
+import {
+  nodeHasChildren,
+  isExactUrlMatch,
+  isDirectory,
+  createNode
+} from "./utils";
+import { getURL } from "./getURL";
+
+import type { Node } from "./types";
+
+const IGNORED_URLS = ["debugger eval code", "XStringBundle"];
+
+/**
+ * Look at the nodes in the source tree, and determine the index of where to
+ * insert a new node. The ordering is index -> folder -> file.
+ * @memberof utils/sources-tree
+ * @static
+ */
+function determineFileSortOrder(
+  nodes: Array<Node>,
+  pathPart: string,
+  isLastPart: boolean,
+  debuggeeUrl: string
+) {
+  const partIsDir = !isLastPart || pathPart.indexOf(".") === -1;
+
+  return nodes.findIndex(node => {
+    const nodeIsDir = nodeHasChildren(node);
+
+    // The index will always be the first thing, so this pathPart will be
+    // after it.
+    if (node.name === "(index)") {
+      return false;
+    }
+
+    // Directory or not, checking root url must be done first
+    if (debuggeeUrl) {
+      const rootUrlMatch = isExactUrlMatch(pathPart, debuggeeUrl);
+      const nodeUrlMatch = isExactUrlMatch(node.name, debuggeeUrl);
+      if (rootUrlMatch) {
+        // pathPart matches root url and must go first
+        return true;
+      }
+      if (nodeUrlMatch) {
+        // Examined item matches root url and must go first
+        return false;
+      }
+      // If neither is the case, continue to compare alphabetically
+    }
+
+    // If both the pathPart and node are the same type, then compare them
+    // alphabetically.
+    if (partIsDir === nodeIsDir) {
+      return node.name.localeCompare(pathPart) >= 0;
+    }
+
+    // If the pathPart and node differ, then stop here if the pathPart is a
+    // directory. Keep on searching if the part is a file, as it needs to be
+    // placed after the directories.
+    return partIsDir;
+  });
+}
+
+/**
+ * @memberof utils/sources-tree
+ * @static
+ */
+export function addToTree(
+  tree: Node,
+  source: SourceRecord,
+  debuggeeUrl: string
+) {
+  const url = getURL(source.get("url"));
+
+  if (
+    IGNORED_URLS.indexOf(url) != -1 ||
+    !source.get("url") ||
+    !url.group ||
+    isPretty(source.toJS())
+  ) {
+    return;
+  }
+
+  url.path = decodeURIComponent(url.path);
+
+  const parts = url.path.split("/").filter(p => p !== "");
+  const isDir = isDirectory(url);
+  parts.unshift(url.group);
+
+  let path = "";
+  let subtree = tree;
+
+  for (let i = 0; i < parts.length; i++) {
+    const part = parts[i];
+    const isLastPart = i === parts.length - 1;
+
+    // Currently we assume that we are descending into a node with
+    // children. This will fail if a path has a directory named the
+    // same as another file, like `foo/bar.js/file.js`.
+    //
+    // TODO: Be smarter about this, which we'll probably do when we
+    // are smarter about folders and collapsing empty ones.
+
+    if (!nodeHasChildren(subtree)) {
+      return;
+    }
+
+    const children = subtree.contents;
+
+    let index = determineFileSortOrder(
+      children,
+      part,
+      isLastPart,
+      i === 0 ? debuggeeUrl : ""
+    );
+
+    const child = children.find(c => c.name === part);
+    if (child) {
+      // A node with the same name already exists, simply traverse
+      // into it.
+      subtree = child;
+    } else {
+      // No node with this name exists, so insert a new one in the
+      // place that is alphabetically sorted.
+      const node = createNode(part, `${path}/${part}`, []);
+      const where = index === -1 ? children.length : index;
+      children.splice(where, 0, node);
+      subtree = children[where];
+    }
+
+    // Keep track of the children so we can tag each node with them.
+    path = `${path}/${part}`;
+  }
+
+  // Overwrite the contents of the final node to store the source
+  // there.
+  if (!isDir) {
+    subtree.contents = source;
+  } else if (!subtree.contents.find(c => c.name === "(index)")) {
+    subtree.contents.unshift(createNode("(index)", source.get("url"), source));
+  }
+}
