+++ src/utils/sources-tree/getURL.js
@@ -1,0 +1,73 @@
+import { parse } from "url";
+import merge from "lodash/merge";
+
+function getFilenameFromPath(pathname?: string) {
+  let filename = "";
+  if (pathname) {
+    filename = pathname.substring(pathname.lastIndexOf("/") + 1);
+    // This file does not have a name. Default should be (index).
+    if (filename == "" || !filename.includes(".")) {
+      filename = "(index)";
+    }
+  }
+  return filename;
+}
+
+export function getURL(sourceUrl: string): { path: string, group: string } {
+  const url = sourceUrl;
+  let def = { path: "", group: "", filename: "" };
+  if (!url) {
+    return def;
+  }
+
+  const { pathname, protocol, host, path } = parse(url);
+  const filename = getFilenameFromPath(pathname);
+
+  switch (protocol) {
+    case "javascript:":
+      // Ignore `javascript:` URLs for now
+      return def;
+
+    case "about:":
+      // An about page is a special case
+      return merge(def, {
+        path: "/",
+        group: url,
+        filename: filename
+      });
+
+    case null:
+      if (pathname && pathname.startsWith("/")) {
+        // If it's just a URL like "/foo/bar.js", resolve it to the file
+        // protocol
+        return merge(def, {
+          path: path,
+          group: "file://",
+          filename: filename
+        });
+      } else if (host === null) {
+        // We don't know what group to put this under, and it's a script
+        // with a weird URL. Just group them all under an anonymous group.
+        return merge(def, {
+          path: url,
+          group: "(no domain)",
+          filename: filename
+        });
+      }
+      break;
+
+    case "http:":
+    case "https:":
+      return merge(def, {
+        path: pathname,
+        group: host,
+        filename: filename
+      });
+  }
+
+  return merge(def, {
+    path: path,
+    group: protocol ? `${protocol}//` : "",
+    filename: filename
+  });
+}
