+++ src/utils/sources-tree/index.js
@@ -1,0 +1,20 @@
+/**
+ * Utils for Sources Tree Component
+ * @module utils/sources-tree
+ */
+
+export { formatTree } from "./formatTree";
+export { addToTree } from "./addToTree";
+export { collapseTree } from "./collapseTree";
+export { getDirectories } from "./getDirectories";
+export { createTree } from "./createTree";
+export { getURL } from "./getURL";
+
+export {
+  nodeHasChildren,
+  isExactUrlMatch,
+  isDirectory,
+  createNode,
+  createParentMap,
+  getRelativePath
+} from "./utils";
